Add files via upload
This project deals with the procedure and validation
for implementing a Universal bi-quad filter. MATLAB is used to simulate the circuit behavior of the transfer function. The frequency response and the transfer function with the operating frequency of 1KHz and pole quality factor Qp=1 and 10 are set to analyze both bandpass filter and band notch filter. Then,
the circuit level behavior is done by mapping the component values with the transfer function in LTSPICE. At last, the
working of Universal bi-quad is verified with the help of simulation in MATLAB, LTSPICE, and developed on ALSK pro board.